+++ .gitignore
@@ -1,3 +1,4 @@
 .build
 .venv
 __pycache__
+.ropeproject
